pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = ["easyths"]
[tool.uv]
index-url = "https://pypi.tuna.tsinghua.edu.cn/simple"

[project]
name = "EasyTHS"
version = "0.1.0"
description = "同花顺交易自动化系统"
authors = [
    { name = "noimank", email = "noimank@163.com" }
]
readme = "README.md"
requires-python = ">=3.12,<3.13"
classifiers = [
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.12",
    "License :: OSI Approved :: MIT License",
    "Operating System :: Microsoft :: Windows",
]
keywords = ["trading", "automation", "tonghuashun", "quantitative"]

dependencies = [
    # 核心依赖
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "pydantic>=2.10.0",
    "pydantic-settings>=2.6.0",
    # Windows GUI自动化
    "pywinauto>=0.6.8",
    "pywin32>=308",
    # 异步任务队列
    "redis>=5.2.0",
    "aioredis>=2.0.1",
    # 日志和监控
    "structlog>=24.4.0",
    "colorama>=0.4.6",
    # 配置文件
    "PyYAML>=6.0.2",
    "python-dotenv>=1.0.1",
    # 工具库
    "python-multipart>=0.0.12",
    "httpx>=0.28.0",
    "aiofiles>=24.1.0",
    "pytesseract>=0.3.13",
    "pillow>=12.0.0",
    "pandas>=2.3.3",
    "tabulate>=0.9.0",
    "python-dateutil>=2.8.2",
    "pyperclip>=1.11.0",
    "mss>=10.1.0",
    "toml>=0.10.2",
    "ddddocr>=1.5.6",
    "psutil>=7.1.3",
]

[project.optional-dependencies]
dev = [
    "black>=24.10.0",
    "ruff>=0.8.0",
    "mypy>=1.13.0",
    "pre-commit>=4.0.1",
]

[dependency-groups]
dev = [
    "black>=24.10.0",
    "ruff>=0.8.0",
    "mypy>=1.13.0",
    "pre-commit>=4.0.1",
]

[project.urls]
Homepage = "https://github.com/noimank/easyths"
Repository = "https://github.com/noimank/easyths.git"
Issues = "https://github.com/noimank/easyths/issues"

# Black 代码格式化配置
[tool.black]
line-length = 88
target-version = ['py312', 'py313']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

# Ruff 代码检查和格式化配置
[tool.ruff]
target-version = "py312"
line-length = 88
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "C4", # flake8-comprehensions
    "UP", # pyupgrade
    "PYI", # flake8-pyi
]
ignore = [
    "E501",  # line too long, handled by black
    "B008",  # do not perform function calls in argument defaults
    "C901",  # too complex
]

[tool.ruff.per-file-ignores]
"__init__.py" = ["F401"]

[tool.ruff.isort]
known-third-party = ["fastapi", "pydantic", "uvicorn"]

# MyPy 类型检查配置
[tool.mypy]
python_version = "3.12"
check_untyped_defs = true
disallow_any_generics = true
disallow_incomplete_defs = true
disallow_untyped_defs = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_return_any = true
strict_equality = true

[[tool.mypy.overrides]]
module = [
    "pywinauto.*",
    "win32gui.*",
    "win32api.*",
    "win32con.*",
]
ignore_missing_imports = true
